+++ src/Plugins/Nop.Plugin.Shipping.FixedByWeightByTotal/Models/ShippingByWeightByTotalModel.cs
@@ -17,22 +17,22 @@ public record ShippingByWeightByTotalModel : BaseNopEntityModel
     }
 
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Store")]
-    public int StoreId { get; set; }
+    public long StoreId { get; set; }
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Store")]
     public string StoreName { get; set; }
 
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Warehouse")]
-    public int WarehouseId { get; set; }
+    public long WarehouseId { get; set; }
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Warehouse")]
     public string WarehouseName { get; set; }
 
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Country")]
-    public int CountryId { get; set; }
+    public long CountryId { get; set; }
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.Country")]
     public string CountryName { get; set; }
 
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.StateProvince")]
-    public int StateProvinceId { get; set; }
+    public long StateProvinceId { get; set; }
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.StateProvince")]
     public string StateProvinceName { get; set; }
 
@@ -40,7 +40,7 @@ public record ShippingByWeightByTotalModel : BaseNopEntityModel
     public string Zip { get; set; }
 
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.ShippingMethod")]
-    public int ShippingMethodId { get; set; }
+    public long ShippingMethodId { get; set; }
     [NopResourceDisplayName("Plugins.Shipping.FixedByWeightByTotal.Fields.ShippingMethod")]
     public string ShippingMethodName { get; set; }
